feat: add script for backing up KVM guests and supporting files
This commit adds a new script called kvm-guest-bkp.sh. This script is used to back up KVM guests and their associated files. It sets the PATH for XML locations and guest images, and defines the XML_DIR and GUEST_DIR variables accordingly. The XML_DIR variable represents the directory where the dump of the guest images configurations will be stored, while the GUEST_DIR variable represents the directory where the .img files and other associated files (if available) will be stored.

Additionally, the commit adds new files to the "noter" directory. The "noter2.c" file is a C program that appends the current date and time to a file called "log.txt" and launches Vim to edit the file. It also checks if the current date is the same as the appended banner and either goes to the end of the file or appends an "EOD" banner and the current date accordingly. The "log.txt" file is a text file that contains timestamps and test messages.

These changes were made to improve the backup process for KVM guests and to provide a simple note-taking functionality.