refactor(tcp-rt.c): refactor code to measure response time of TCP connection
The code has been refactored to measure the response time of a TCP connection. The `server_addr` struct is now initialized with the server IP and port. The `gettimeofday` function is used to get the start and end time of the connection. The connection is established using the `connect` function and the response time is calculated by subtracting the start time from the end time. The response time is then printed to the console.